requirements.txt
langchain
langchain-openai
pydantic
langgraph
pyyaml
matplotlib
seaborn
pandas
datasets
